% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{build_knn_graph}
\alias{build_knn_graph}
\title{Build kNN network}
\usage{
build_knn_graph(
  data,
  k = 20,
  dist_type = c("euclidean", "pearson", "raw"),
  use_seurat_snn = TRUE,
  mutual = TRUE,
  jaccard_weighted = TRUE,
  jaccard_prune = 1/15,
  return_igraph = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{The data matrix}

\item{k}{The number of neighbors}

\item{dist_type}{The type of distance. Should be one of "euclidean", "pearson" and "raw"}

\item{use_seurat_snn}{If TRUE, the FindNeighbors function in Seurat will be called to calculate}

\item{mutual}{Only consider mutual nearest neighbors. Ignore if use_seurat_snn is TRUE}

\item{jaccard_weighted}{If TRUE, edges of the resulted graph are weighted by Jaccard index. Ignore if use_seurat_snn is TRUE}

\item{jaccard_prune}{Prune an edge if its weight is smaller than this value. Ignore if use_seurat_snn is FALSE and jaccard_weighted is FALSE}

\item{return_igraph}{if TRUE, an igraph object instead of the adjacent matrix will be returned}

\item{verbose}{If TRUE, progress message is provided}
}
\value{
A Matrix of kNN network adjacant matrix. If return_igraph is TRUE, the corresponding igraph object is returned
}
\description{
Calculate the kNN-network given a data matrix
}
