% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.r
\name{construct_pseudocells}
\alias{construct_pseudocells}
\title{Construct pseudocells by averaging transcriptome of similar cells}
\usage{
construct_pseudocells(
  knn_adj,
  ratio,
  init_dist = 1,
  max_dist = 2,
  min_pooled_cells = 2,
  min_seed_num = 1
)
}
\arguments{
\item{knn_adj}{The adjacent matrix of the kNN network}

\item{ratio}{The downsample rate}

\item{init_dist}{Maximal distance on the kNN netowkr to group a cell to the selected capital or pseudocell center}

\item{max_dist}{A cell which is not grouped to any pseudocell territory is grouped to the closest one if its distance to the capital is no more than max_dist}

\item{min_pooled_cells}{Only pseudocells covering at least this number of cells are valid}

\item{min_seed_num}{The minimum number of seed capitals to start the procedure}
}
\value{
A data.frame indicating indices of cells and pseudocells
}
\description{
This function implements the procedure to average the expression profiles
of similar cells in order to construct pseudocells
}
