% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.r, R/generics.r
\name{css_project.default}
\alias{css_project.default}
\alias{css_project.Seurat}
\alias{css_project}
\title{Prject data to the CSS representation using the given calculation model}
\usage{
\method{css_project}{default}(object, model)

\method{css_project}{Seurat}(
  object,
  model,
  reduction.name = "css_proj",
  reduction.key = "CSSPROJ_"
)

css_project(object, ...)
}
\arguments{
\item{object}{An object}

\item{model}{Calculation model of the reference CSS representation}

\item{reduction.name}{Reduction name of the projected CSS representation in the returned Seurat object}

\item{reduction.key}{Reduction key of the projected CSS representation in the returned Seurat object}
}
\description{
Calculate CSS representation using a given model. It is used to project
new data to the existing reference atlas which is already represented by CSS
}
