% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.r
\name{estimate_projection_failure}
\alias{estimate_projection_failure}
\title{Estimate likelihood of projection failure}
\usage{
estimate_projection_failure(
  data_ref,
  data_query,
  k = 50,
  seed = NULL,
  do_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data_ref}{Reference data matrix}

\item{data_query}{Query data matrix}

\item{k}{Number of neighbors}

\item{seed}{The seed value for the Gaussian mixture model estimate}

\item{do_plot}{Whether to plot the fitted mixture model (as density plot)}

\item{verbose}{Whether to output the verbose message}
}
\value{
A data frame with normalized projection distance and estimated projection failure likelihood
}
\description{
Calculate the normalized distances of each query sample to the projected reference subset,
and use a Gaussian mixture model to estimate the likelihood of projection failure, which is
featured by large distance to the reference
}
