% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rank_input_matrix}
\alias{rank_input_matrix}
\title{Generate ranked sparse matrix}
\usage{
rank_input_matrix(mat)
}
\arguments{
\item{mat}{The input matrix, can be dense or sparse}
}
\value{
A matrix, whose dense/sparse is the same as the input, with each column ranked
}
\description{
This function convert a non-negative matrix into a column-ranked matrix
}
