% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/rss.r
\name{ref_sim_spectrum}
\alias{ref_sim_spectrum}
\alias{ref_sim_spectrum.default}
\alias{ref_sim_spectrum.Seurat}
\title{Calculate Reference Similarity Spectrum}
\usage{
ref_sim_spectrum(object, ...)

\method{ref_sim_spectrum}{default}(
  object,
  ref,
  method = c("pearson", "spearman"),
  use_fast_rank = TRUE,
  scale = TRUE
)

\method{ref_sim_spectrum}{Seurat}(
  object,
  ref,
  ...,
  as_assay = FALSE,
  assay.name = "rss",
  reduction.name = "rss",
  reduction.key = "RSS_"
)
}
\arguments{
\item{object}{An object}

\item{ref}{Expression matrix of the reference data set}

\item{method}{Correlation method}

\item{use_fast_rank}{When the presto package is available, use its rank_matrix function to rank sparse matrix}

\item{scale}{If TRUE, z-transform is applied to the calculated similarities across reference samples}

\item{as_assay}{When it is TRUE, the output is returned as an Assay object in the Seurat object}

\item{assay.name}{Assay name of the RSS representation in the returned Seurat object}

\item{reduction.name}{Reduction name of the RSS representation in the returned Seurat object}

\item{reduction.key}{Reduction key of the RSS representation in the returned Seurat object}
}
\description{
Calculate Reference Similarity Spectrum (RSS), given expression of the
data and expression of the external reference data set.
}
