% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/regress.r
\name{regress_out_from_embeddings}
\alias{regress_out_from_embeddings}
\alias{regress_out_from_embeddings.default}
\alias{regress_out_from_embeddings.Seurat}
\title{Regress out variables from cell embeddings}
\usage{
regress_out_from_embeddings(object, ...)

\method{regress_out_from_embeddings}{default}(object, vars)

\method{regress_out_from_embeddings}{Seurat}(
  object,
  reduction,
  vars_to_regress,
  reduction.name = reduction,
  reduction.key = NULL
)
}
\arguments{
\item{object}{An object}

\item{vars}{Data frame of the variables to regress out}

\item{reduction}{Name of the reduction object to use}

\item{vars_to_regress}{Variables in the meta.data slot for regressing out}

\item{reduction.name}{Name of the new reduction object}

\item{reduction, key}{Key of the returned reduction}
}
\description{
Regress out variables from cell embeddings with a linear model
}
