% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rowNorm}
\alias{rowNorm}
\title{Normalize each row of the given matrix to be sum of 1}
\usage{
rowNorm(mat)
}
\arguments{
\item{mat}{The input matrix}
}
\value{
The normalized matrix
}
\description{
This function normalizes the given matrix in a row-wise manner so that
the sum of every row is one
}
