% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/pca.r
\name{run_PCA}
\alias{run_PCA}
\alias{run_PCA.default}
\alias{run_PCA.Seurat}
\title{Principal component analysis on dimension reduced data}
\usage{
run_PCA(object, ...)

\method{run_PCA}{default}(object, npcs = 50, return_rotation = F)

\method{run_PCA}{Seurat}(
  object,
  reduction,
  npcs = 10,
  reduction.name = NULL,
  reduction.key = NULL
)
}
\arguments{
\item{object}{An object}

\item{npcs}{Number of PCs to compute}

\item{return_rotation}{Whether to return the rotation matrix}

\item{reduction}{The reduction to run PCA on}

\item{reduction.name}{Name of the returned reduction}

\item{reduction.key}{Key of the returned reduction}
}
\description{
Do PCA analysis on dimension reduced data to further reduce
data dimensionality
}
