% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.r
\name{transfer_labels}
\alias{transfer_labels}
\title{Transfer labels from the reference to the query data set}
\usage{
transfer_labels(
  data_ref = NULL,
  data_query = NULL,
  knn_ref_query = NULL,
  label_ref,
  k = 50,
  thres_prop_match = 0.5
)
}
\arguments{
\item{data_ref}{Reference data matrix}

\item{data_query}{Query data matrix}

\item{knn_ref_query}{KNN matrix, where each row represents one sample in the query data, and each column represents one nearest neighbors in the reference data. If NULL, it is calculated using data_ref and data_query}

\item{label_ref}{Class labels of each sample in the reference}

\item{k}{Number of neighbors for classification}

\item{thres_prop_match}{If no class gets higher supported proportion than this value, the prediction fails for the query sample and NA is returned}
}
\value{
A vector of the predicted/transferred labels of the query data
}
\description{
A kNN classifier is used to predict labels of the query data, given the reference
data with the same data representation
}
